Relocate resource classes
