+++ lib/registryBuilder.js
@@ -5,9 +5,9 @@ var path = require('path');
 
 var Registry = require('./registry');
 
-var Document = require('./document');
-var Category = require('./category');
-var CustomPage = require('./customPage');
+var Document = require('./resources/document');
+var Category = require('./resources/category');
+var CustomPage = require('./resources/customPage');
 
 var RegistryBuilder = {};
