+++ lib/requestor.js
@@ -196,6 +196,15 @@ Requestor.uploadBodyFns = {
     }
 };
 
+Requestor.isCustomSlug = function(resource, responseBody) {
+    var slugSet = !!resource.slug;
+    var isCustomSlug = slugSet && responseBody.slug !== resource.slug;
+    var isDoc = !!resource.categorySlug;
+    var wasPost = resource.method === 'post';
+
+    return isCustomSlug && isDoc && wasPost;
+};
+
 Requestor.prototype.uploadReqSettings = function(resource) {
     var reqSettings = {};
     var urlGen = new UrlGenerator(this.projectName, resource.version);
@@ -211,15 +220,6 @@ Requestor.prototype.uploadReqSettings = function(resource) {
     reqSettings.body = Requestor.uploadBodyFns[resource.resourceType](resource);
 
     return reqSettings;
-};
-
-Requestor.isCustomSlug = function(resource, responseBody) {
-    var slugSet = !!resource.slug;
-    var isCustomSlug = slugSet && responseBody.slug !== resource.slug;
-    var isDoc = !!resource.categorySlug;
-    var wasPost = resource.method === 'post';
-
-    return isCustomSlug && isDoc && wasPost;
 };
 
 Requestor.prototype.setCustomSlug = function(resource, responseBody, cb) {
@@ -282,7 +282,7 @@ Requestor.prototype.resourceToUploadRequestFn = function(resource) {
 
 // Uploads an array of resources
 Requestor.prototype.uploadResources = function(resources, cb) {
-    var requestFns = resources.map(this.resourceToUploadRequestFn);
+    var requestFns = resources.map(this.resourceToUploadRequestFn.bind(this));
 
     async.parallelLimit(requestFns, MAX_PARALLEL_REQUESTS, function(err, results) {
         if (err) {
