+++ lib/resources/category.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('./utils');
+var utils = require('../utils');
 var path = require('path');
 
 var CATEGORY_PROPERTIES = ['title', 'slug', 'excerpt', 'body', 'order', 'pages'];
