+++ lib/resources/customPage.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var path = require('path');
-var utils = require('./utils');
+var utils = require('../utils');
 
 var CUSTOM_PAGE_PROPERTIES = ['title', 'html', 'slug'];
