+++ lib/resources/document.js
@@ -2,7 +2,7 @@
 
 var fs = require('fs');
 var path = require('path');
-var utils = require('./utils');
+var utils = require('../utils');
 
 var DOC_PROPERTIES = ['title', 'slug', 'excerpt', 'body', 'order'];
